phonebook update

+++ src/components/ContactList/ContactList.js
@@ -13,22 +13,27 @@ export class ContactList extends Component {
   };
 
   render() {
-    return (
-      <div className={styles.list}>
-        {this.props.contactList.map((item) => (
-          <li key={item.id}>
-            <span>{item.name}</span>
-            <span>{item.number}</span>
-            <button
-              className={styles.button}
-              onClick={() => this.deleteItem(item.id)}
-            >
-              Delete
-            </button>
-          </li>
-        ))}
-      </div>
-    );
+    // console.log(this.props.contactList);
+    if (this.props.contactList) {
+      return (
+        <div className={styles.list}>
+          {this.props.contactList.map((item) => (
+            <li key={item.id}>
+              <span>{item.name}</span>
+              <span>{item.number}</span>
+              <button
+                className={styles.button}
+                onClick={() => this.deleteItem(item.id)}
+              >
+                Delete
+              </button>
+            </li>
+          ))}
+        </div>
+      );
+    } else {
+      return null;
+    }
   }
 }
